Define maps to transform a corner coordinate to canonical points
There are two that are potentially of interest. The canonical point that
is *inside* the white area the corner borders, and the point immediately
on the border *outside* this area.

If we assume that all borders are 1px wide, then corners that are
actually adjacent will have the same outside coordinate.